refactor(feature-020): C5 prep - hoist the binary dispatcher into cli::entry; main.rs becomes a shim
Behavior-preserving preparation for the T031 exposure flip: the flip
retires the raw symforge::* module surface the binary consumed, so the
whole dispatcher (startup planning, daemon/serve/stdio runners, periodic
checkpoint spawn, and its tests) moves from src/main.rs into the crate
as cli::entry, with symforge:: paths rewritten to crate::. main.rs
shrinks to an 8-line shim calling cli::entry::run_main; the flip commit
rewires it through the introduced server_api::run door.

Follow-through recorded honestly:
- The stdio bootstrap edges (activate_surface, admit_project, observer
  capture) moved with the code, so src/cli/entry.rs replaces src/main.rs
  in WIRED_PRODUCTION_FILES (anti-vacuity holds on the new row; the shim
  holds no lifecycle edge).
- The frozen callbacks census anchors spawn_periodic_checkpoint and the
  background_verify spawn at src/main.rs; the constructs now live in
  cli/entry.rs, adding their mechanical shadow pairs to the validator's
  enumerated mid-cut lines (ten during the C5 window, documented in the
  execution map; the whole preactivation branch retires at the flip).
- tests/session_cache_hit.rs dogfoods this repository and fetched
  src/main.rs::main as its 'substantive body'; the shim's main is 128
  bytes - smaller than the cache-hit banner - so the fixture now fetches
  cli/entry.rs::run_main, preserving the test's intent.

Mid-battery infrastructure note (not a code defect): the first two suite
attempts died to a Windows commit-limit exhaustion (0xc000012d) during
16-way compilation, corrupting target/ artifacts (E0786 cascades, one
201s ICE). Recovery per the documented ladder ended in a full cargo
clean (206.8 GiB removed) and a -j 8 rebuild, which passed.

Gates: fmt clean; dark seal 9/9 (full-source pin refreshed post-fmt:
195 files - entry.rs joins the census); clippy --all-targets -D warnings
exit 0; embed gate 1336 passed; full serial suite exit 0 (579s, lib 3272
passed incl. the moved dispatcher tests); validator at the ten
enumerated mid-cut lines.
